crudcast/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
pymongo==3.7.2
PyYAML==3.13
Werkzeug==0.14.1
flasgger==0.9.2